pyproject.toml
[tool.poetry]
name = "inky-subway"
version = "0.1.0"
description = ""
authors = ["Chris Barry <me@chrisb.xyz>"]

[tool.poetry.dependencies]
python = "^3.9"
underground = "^0.3.0"
inky = "^1.0.2"
Pillow = "^8.0.1"
dotenv = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
